+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+##### 0.9.1 - xx May 2014
+
+###### Backwards compatible bug fixes
+- #68 - Async methods should honor methods on a resource definition.
+- #69 - Failed requests should throw an error, not just return it
+- #70 - Make `params` and `params.query` optional
+
 ##### 0.9.0 - 22 May 2014
 
 ###### Breaking API changes